refactor: switch to kebab-case naming
