+++ src/pages/_app.tsx
@@ -4,7 +4,7 @@ import NextHead from "next/head";
 import { type AppProps } from "next/app";
 import { type NextPage } from "next";
 import { DefaultSeo, type DefaultSeoProps } from "next-seo";
-import { GlobalStyles } from "~/styles/GlobalStyles";
+import { GlobalStyles } from "~/styles/global-styles";
 
 export type NextPageWithLayout<PageProps = {}> = NextPage<PageProps> & {
   getLayout?: (page: React.ReactElement<PageProps>) => React.ReactNode;
